Added custom hook to delete qr using react-query
