+++ src/hooks/useAPI/useQuoteRequests.jsx
@@ -1,5 +1,9 @@
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { fetchAllQuoteRequests, updateQuoteRequestById } from '../../services/api/quote_request';
+import {
+  deleteQuoteRequestById,
+  fetchAllQuoteRequests,
+  updateQuoteRequestById
+} from '../../services/api/quote_request';
 
 // Custom react-query hook for quote requests
 export const useFetchAllQuoteRequests = () => {
@@ -9,6 +13,34 @@ export const useFetchAllQuoteRequests = () => {
   });
 
   return { data, isLoading, isError };
+};
+
+// Custom react-query hook for deleting a qr by id
+export const useDeleteQRById = (toast) => {
+  const queryClient = useQueryClient();
+  return useMutation((itemId) => deleteQuoteRequestById(itemId), {
+    onError: (error) => {
+      toast({
+        position: `top`,
+        title: `Error occured deleting QR!`,
+        description: `Error: ${error.message}`,
+        status: 'error',
+        duration: 5000,
+        isClosable: true
+      });
+    },
+    onSuccess: (data) => {
+      queryClient.invalidateQueries({ queryKey: ['quoteRequests'] });
+      toast({
+        position: `top`,
+        title: `QR #${data} deleted!`,
+        description: `We've deleted QR #${data} for you succesfully! 🚀`,
+        status: 'success',
+        duration: 5000,
+        isClosable: true
+      });
+    }
+  });
 };
 
 // Custom react-query hook for updating a qr by id
